Internals

Base.read

Base.read

Base.read

Base.read

Base.read

Base.read

Base.read

Base.read

Base.read

Base.read

Makie.plot

OceanRobots.CCHDO.ancillary_files

OceanRobots.CCHDO.download

OceanRobots.CCHDO.extract_json_table

OceanRobots.GDP.download

OceanRobots.GDP.list_files

OceanRobots.GDP_CloudDrift.add_ID!

OceanRobots.GDP_CloudDrift.add_index!

OceanRobots.GDP_CloudDrift.region_subset

OceanRobots.GDP_CloudDrift.to_DataFrame

OceanRobots.GDP_CloudDrift.to_Grid

OceanRobots.GDP_CloudDrift.trajectory_stats

OceanRobots.Glider_EGO_module.file_lists

OceanRobots.NOAA.download

OceanRobots.NOAA.download

OceanRobots.NOAA.download

OceanRobots.NOAA.download_historical_txt

OceanRobots.NOAA.list_realtime

OceanRobots.NOAA.list_stations

OceanRobots.NOAA.read_historical_monthly

OceanRobots.NOAA.read_historical_nc

OceanRobots.NOAA.read_historical_txt

OceanRobots.NOAA.read_station

OceanRobots.OceanOPS.get_list

OceanRobots.OceanOPS.get_list_pos

OceanRobots.OceanOPS.get_platform

OceanRobots.OceanOPS.get_url

OceanRobots.OceanOPS.list_platform_types

OceanRobots.OceanSites.index

OceanRobots.OceanSites.read_WHOTS

OceanRobots.OceanSites.read_variables

OceanRobots.query

OceanRobots.OceanOPS.get_list — Function

get_list(nam=:Argo; status="OPERATIONAL")

Get list of platform IDs from OceanOPS API.

For more information see

https://www.ocean-ops.org/api/1/help/

https://www.ocean-ops.org/api/1/help/?param=platformstatus

list_Argo1=OceanOPS.get_list(:Argo,status="OPERATIONAL")
list_Argo2=OceanOPS.get_list(:Argo,status="CONFIRMED")
list_Argo3=OceanOPS.get_list(:Argo,status="REGISTERED")
list_Argo4=OceanOPS.get_list(:Argo,status="INACTIVE")

OceanRobots.OceanOPS.get_list_pos — Function

get_list_pos(nam=:Argo; status="OPERATIONAL")

Get list of platform positions from OceanOPS API.

For more information see

https://www.ocean-ops.org/api/1/help/

https://www.ocean-ops.org/api/1/help/?param=platformstatus

OceanRobots.OceanOPS.get_platform — Method

get_platform(i)

Get info on platform with id=i (e.g., float or drifter) from OceanOPS API.

For more information see https://www.ocean-ops.org/api/1/help/

list_Drifter=OceanOPS.get_list(:Drifter)
tmp=OceanOPS.get_platform(list_Drifter[1000])

OceanRobots.OceanOPS.get_url — Function

get_url(nam=:Argo; status="OPERATIONAL")

API/GET URL to OceanOPS API that will list platforms of chosen type.

Two URLs are reported; the second includes platform positions.

For more information see

https://www.ocean-ops.org/api/1/help/

https://www.ocean-ops.org/api/1/help/?param=platformstatus

https://www.ocean-ops.org/api/1/help/?param=platformtype

OceanRobots.OceanOPS.list_platform_types — Method

list_platform_types()

List platform types.

Base.read — Method

read(x::SurfaceDrifter,ii::Int)

Open file number ii from NOAA ftp server using NCDatasets.jl.

Server : ftp://ftp.aoml.noaa.gov/pub/phod/lumpkin/hourly/v2.00/netcdf/

Note: the first time this method is used, it calls GDP.list_files() to get the list of drifters from server, and save it to a temporary file.

using OceanRobots
sd=read(SurfaceDrifter(),1)

Base.read — Method

read(x::SurfaceDrifter; ID=300234065515480, version="v2.01")

Download file from NOAA http server read it using NCDatasets.jl.

Server : https://www.aoml.noaa.gov/ftp/pub/phod/lumpkin/hourly/

using OceanRobots
sd=read(SurfaceDrifter(),ID=300234065515480)

OceanRobots.GDP.download — Function

download(list_files,ii=1)

Download one drifter file from NOAA ftp server.

ftp://ftp.aoml.noaa.gov/pub/phod/lumpkin/hourly/v2.00/netcdf/

list_files=GDP.list_files()
fil=GDP.download(list_files,1)

OceanRobots.GDP.list_files — Method

list_files()

Get list of drifter files from NOAA ftp server or the corresponding webpage.

ftp://ftp.aoml.noaa.gov/pub/phod/lumpkin/hourly/v2.00/netcdf/

https://www.aoml.noaa.gov/ftp/pub/phod/lumpkin/hourly/v2.00/netcdf/

Base.read — Method

read(x::CloudDrift, file)

Read a GDP/CloudDrift file.

OceanRobots.GDP_CloudDrift.add_ID! — Method

add_ID!(df,ds)

OceanRobots.GDP_CloudDrift.add_index! — Method

add_index!(df)

OceanRobots.GDP_CloudDrift.region_subset — Method

region_subset(df,lons,lats,dates)

Subset of df that's within specified date and position ranges.

OceanRobots.GDP_CloudDrift.to_DataFrame — Method

to_DataFrame(ds)

OceanRobots.GDP_CloudDrift.to_Grid — Method

to_Grid(gdf,grid)

OceanRobots.GDP_CloudDrift.trajectory_stats — Method

trajectory_stats(gdf)

Base.read — Method

read(x::NOAAbuoy,args...)

Read a NOAA buoy file (past month).

Base.read — Method

read(x::NOAAbuoy_monthly,args...)

Read a NOAA buoy file (historical).

OceanRobots.NOAA.download — Function

NOAA.download(stations::Union(Array,Int),path=tempdir())

Download files listed in stations from ndbc.noaa.gov to path.

OceanRobots.NOAA.download — Function

NOAA.download(sta::String,path=tempdir())

Download files for stations sta from ndbc.noaa.gov to path.

OceanRobots.NOAA.download — Method

NOAA.download(MC::ModelConfig)

Download files listed in MC.inputs["stations"] from ndbc.noaa.gov to pathof(MC).

OceanRobots.NOAA.download_historical_txt — Method

NOAA.download_historical_txt(ID,years)

Download files from https://www.ndbc.noaa.gov to temporary folder for chosen float ID and years.

OceanRobots.NOAA.list_realtime — Method

NOAA.list_realtime(;ext=:all)

Get either files list from https://www.ndbc.noaa.gov/data/realtime2/ or list of buoy codes that provide some file type (e.g. "txt" for "Standard Meteorological Data")

lst0=NOAA.list_realtime()
lst1=NOAA.list_realtime(ext=:txt)

OceanRobots.NOAA.list_stations — Method

NOAA.list_stations()

Get stations list from https://www.ndbc.noaa.gov/to_station.shtml

OceanRobots.NOAA.read_historical_monthly — Function

NOAA.read_historical_monthly(ID,years)

Read files from https://www.ndbc.noaa.gov to temporary folder for chosen float ID and year y.

OceanRobots.NOAA.read_historical_nc — Method

NOAA.read_historical_nc(ID,year)

Read files from https://www.ndbc.noaa.gov to temporary folder for chosen float ID and year y.

OceanRobots.NOAA.read_historical_txt — Method

NOAA.read_historical_txt(ID,y)

Read files from https://www.ndbc.noaa.gov to temporary folder for chosen float ID and year y.

OceanRobots.NOAA.read_station — Function

NOAA.read_station(station,path=tempdir())

Read station file from specified path, and add meta-data (units and descriptions).

Base.read — Function

read(x::ShipCruise, ID="unknown")

Read ShipCruise data.

OceanRobots.CCHDO.ancillary_files — Method

ancillary_files(cruise::Union{Symbol,String})

using OceanRobots
ID="33RR20230722"
list=OceanRobots.CCHDO.ancillary_files(ID)

OceanRobots.CCHDO.download — Function

CCHDO.download(cruise::Union(Symbol,Symbol[]),path=tempdir())

Download files listed in stations from cchdo.ucsd.edu/cruise/ to path.

using OceanRobots
ID="33RR20160208"
path=OceanRobots.CCHDO.download(ID)

OceanRobots.CCHDO.extract_json_table — Function

extract_json_table(url)

using JSON3, HTTP

url="https://cchdo.ucsd.edu/search?q=GO-SHIP"
url="https://cchdo.ucsd.edu/search?bbox=-75,-60,20,65" #Atlantic
url="https://cchdo.ucsd.edu/search?dtstart=1999-12-31&dtend=2000-01-31" #one month
url="https://cchdo.ucsd.edu/search?q=ARS01" # BATS
url="https://cchdo.ucsd.edu/search?q=PRS02" #HOT

table=extract_json_table(url)

ta=table[1]
xy1=ta.track.coordinates
x1=[i[1] for i in xy1]
y1=[i[2] for i in xy1]  

using GLMakie
scatter(x1,y1)

Base.read — Function

read(x::Glider_Spray, file::String)

Read a Spray Glider file.

Base.read — Function

read(x::Glider_EGO, ID=1)

Read a EGO Glider files.

OceanRobots.Glider_EGO_module.file_lists — Function

file_lists(k=1:2)

Get list of EGO glider files from ftp server ftp://ftp.ifremer.fr/ifremer/glider/v2/

missions,folders,files=Glider_EGO_module.file_lists(1:10)
Glider_EGO_module.glider_download(files[1][1])

data=read(Glider_EGO(),2)

fig_glider=plot(data)

ds=data.ds,variable="CHLA")

Base.read — Function

read(x::OceanSite, ID=:WHOTS)

Read OceanSite data.

OceanRobots.OceanSites.index — Method

index()

Download, read and process the oceansites_index.txt file. Return a DataFrame.

oceansites_index=OceanSites.index()

OceanRobots.OceanSites.read_WHOTS — Function

read_WHOTS(fil)

Read an WHOTS file.

file="DATA_GRIDDED/WHOTS/OS_WHOTS_200408-201809_D_MLTS-1H.nc"
data,units=OceanSites.read_WHOTS(file)

OceanRobots.OceanSites.read_variables — Method

read_variables(file,args...)

Open file from opendap server.

file="DATA_GRIDDED/WHOTS/OS_WHOTS_200408-201809_D_MLTS-1H.nc"
OceanSites.read_variables(file,:lon,:lat,:time,:TEMP)