Reference Manual

OceanRobots.ArgoFiles.download

OceanRobots.ArgoFiles.read

OceanRobots.ArgoFiles.scan_txt

OceanRobots.GDP.download

OceanRobots.GDP.list_files

OceanRobots.GDP.read

OceanRobots.GliderFiles.read

OceanRobots.NOAA.download

OceanRobots.NOAA.download_historical_txt

OceanRobots.NOAA.read

OceanRobots.NOAA.read_historical_monthly

OceanRobots.NOAA.read_historical_nc

OceanRobots.NOAA.read_historical_txt

OceanRobots.OceanOPS.csv_listings

OceanRobots.OceanOPS.get_list

OceanRobots.OceanOPS.get_list_pos

OceanRobots.OceanOPS.get_platform

OceanRobots.OceanOPS.get_table

OceanRobots.OceanOPS.get_url

OceanRobots.OceanSites.index

OceanRobots.OceanSites.read

OceanRobots.OceanSites.read_WHOTS

OceanRobots.THREDDS.parse_catalog

OceanRobots.THREDDS.parse_catalog_NOAA_buoy

OceanRobots.cmems_sla.subset

OceanRobots.podaac_sla.subset

OceanRobots.OceanOPS.csv_listings — Method

csv_listings()

List csv files available on the https://www.ocean-ops.org/share/ server.

OceanRobots.OceanOPS.get_list — Function

get_list(nam=:Argo; status="OPERATIONAL")

Get list of platform IDs from OceanOPS API.

For more information see

https://www.ocean-ops.org/api/1/help/

https://www.ocean-ops.org/api/1/help/?param=platformstatus

list_Argo1=OceanOPS.get_list(:Argo,status="OPERATIONAL")
list_Argo2=OceanOPS.get_list(:Argo,status="CONFIRMED")
list_Argo3=OceanOPS.get_list(:Argo,status="REGISTERED")
list_Argo4=OceanOPS.get_list(:Argo,status="INACTIVE")

OceanRobots.OceanOPS.get_list_pos — Function

get_list_pos(nam=:Argo; status="OPERATIONAL")

Get list of platform positions from OceanOPS API.

For more information see

https://www.ocean-ops.org/api/1/help/

https://www.ocean-ops.org/api/1/help/?param=platformstatus

OceanRobots.OceanOPS.get_platform — Method

get_platform(i)

Get info on platform with id=i (e.g., float or drifter) from OceanOPS API.

For more information see https://www.ocean-ops.org/api/1/help/

list_Drifter=OceanOPS.get_list(:Drifter)
tmp=OceanOPS.get_platform(list_Drifter[1000])

OceanRobots.OceanOPS.get_table — Function

get_table(s::Symbol,i=1)

Read the csv_listings()[s][i] table. Download file if needed.

using OceanRobots
tab_Argo=OceanOPS.get_table(:Argo,1)

OceanRobots.OceanOPS.get_url — Function

get_url(nam=:Argo; status="OPERATIONAL")

API/GET URL to OceanOPS API that will list platforms of chosen type.

Two URLs are reported; the second includes platform positions.

For more information see

https://www.ocean-ops.org/api/1/help/

https://www.ocean-ops.org/api/1/help/?param=platformstatus

https://www.ocean-ops.org/api/1/help/?param=platformtype

OceanRobots.GDP.download — Function

download(list_files,ii=1)

Download one drifter file from NOAA ftp server.

ftp://ftp.aoml.noaa.gov/pub/phod/lumpkin/hourly/v2.00/netcdf/

list_files=GDP.list_files()
fil=GDP.download(list_files,1)

OceanRobots.GDP.list_files — Method

list_files()

Get list of drifter files from NOAA ftp server or the corresponding webpage.

ftp://ftp.aoml.noaa.gov/pub/phod/lumpkin/hourly/v2.00/netcdf/

https://www.aoml.noaa.gov/ftp/pub/phod/lumpkin/hourly/v2.00/netcdf/

OceanRobots.GDP.read — Method

read(filename::String)

Open file from NOAA ftp server using NCDatasets.Dataset.

ftp://ftp.aoml.noaa.gov/pub/phod/lumpkin/hourly/v2.00/netcdf/ or the corresponding webpage

list_files=GDP.list_files()
fil=GDP.download(list_files,1)
ds=GDP.read(fil)

OceanRobots.NOAA.download — Method

NOAA.download(MC::ModelConfig)

Download files listed in MC.inputs["stations"] from ndbc.noaa.gov to pathof(MC).

OceanRobots.NOAA.download_historical_txt — Method

NOAA.download_historical_txt(ID,years)

Download files from https://www.ndbc.noaa.gov to temporary folder for chosen float ID and years.

OceanRobots.NOAA.read — Method

NOAA.read(MC,sta)

Read station sta file from pathof(MC). Meta-data is provided in NOAA.units and NOAA.descriptions.

OceanRobots.NOAA.read_historical_monthly — Function

NOAA.read_historical_monthly(ID,years)

Read files from https://www.ndbc.noaa.gov to temporary folder for chosen float ID and year y.

OceanRobots.NOAA.read_historical_nc — Method

NOAA.read_historical_nc(ID,year)

Read files from https://www.ndbc.noaa.gov to temporary folder for chosen float ID and year y.

OceanRobots.NOAA.read_historical_txt — Method

NOAA.read_historical_txt(ID,y)

Read files from https://www.ndbc.noaa.gov to temporary folder for chosen float ID and year y.

OceanRobots.GliderFiles.read — Method

GliderFiles.read(file::String)

Read a Spray Glider file.

OceanRobots.ArgoFiles.download — Method

ArgoFiles.download(files_list,wmo)

Download an Argo profiler file.

OceanRobots.ArgoFiles.read — Method

ArgoFiles.read(fil)

Read an Argo profiler file.

OceanRobots.ArgoFiles.scan_txt — Function

ArgoFiles.scan_txt(fil="ar_index_global_prof.txt"; do_write=false)

Scan the Argo file lists and return summary tables in DataFrame format. Write to csv file if istrue(do_write).

ArgoFiles.scan_txt("ar_index_global_prof.txt",do_write=true)
ArgoFiles.scan_txt("argo_synthetic-profile_index.txt",do_write=true)

OceanRobots.OceanSites.index — Method

index()

Download, read and process the oceansites_index.txt file. Return a DataFrame.

oceansites_index=OceanSites.index()

OceanRobots.OceanSites.read — Method

read(file,args...)

Open file from opendap server.

file="DATA_GRIDDED/WHOTS/OS_WHOTS_200408-201809_D_MLTS-1H.nc"
OceanSites.read(file,:lon,:lat,:time,:TEMP)

OceanRobots.OceanSites.read_WHOTS — Function

read_WHOTS(fil)

Read an WHOTS file.

file="DATA_GRIDDED/WHOTS/OS_WHOTS_200408-201809_D_MLTS-1H.nc"
data,units=OceanSites.read_WHOTS(file)

OceanRobots.THREDDS.parse_catalog — Function

parse_catalog(url,recursive=true)

Starting from an xml (not html) thredds catalog look for both subfolders and files; they are identified based on the href and urlPath attributes respectively. If recursive is set to true then go down in subfolders and do the same until only files are found; in this case the returned folders should be empty and files can be extensive.

For more on thredds servers, see https://www.unidata.ucar.edu/software/tds/current/catalog/.

url="https://dods.ndbc.noaa.gov/thredds/catalog/oceansites/long_timeseries/WHOTS/catalog.xml"
#url="https://dods.ndbc.noaa.gov/thredds/catalog/data/catalog.xml"
#url="https://dods.ndbc.noaa.gov/thredds/catalog/oceansites-tao/catalog.xml"
#url="https://dods.ndbc.noaa.gov/thredds/catalog/tao-ctd/catalog.xml"
#url="https://dods.ndbc.noaa.gov/thredds/catalog/hfradar/catalog.xml"
#url="https://dods.ndbc.noaa.gov/thredds/catalog/oceansites/catalog.xml"

files,folders=parse_catalog(url)

OceanRobots.THREDDS.parse_catalog_NOAA_buoy — Function

parse_catalog_NOAA_buoy(ID=44013)

Use parse_catalog_NOAA_buoy to build files_year,files_url lists for buoy ID.

OceanRobots.podaac_sla.subset — Method

podaac_sla.subset()

For download directions, see this site

podaac_sla.subset()

OceanRobots.cmems_sla.subset — Method

cmems_sla.subset()

For download directions, see this site

For data documentation, see this page

cmems_sla.subset(username=username,password=password)