Reference Manual

API

MakieCore.plot — Function

plot(x::SurfaceDrifter)

using OceanRobots
lst=GDP.list_files()
drifter=read(SurfaceDrifter(),1,list_files=lst)
plot(drifter)

plot(x::OceanSite,args...)

using OceanRobots, Dates
whots=read(OceanSite(),:WHOTS)
plot(whots,DateTime(2005,1,1),DateTime(2005,2,1))

plot(x::NOAAbuoy,var)

using OceanRobots
buoy=read(NOAAbuoy(),41046)
plot(buoy,"PRES")

plot(x::NOAAbuoy_monthly, var=""; option=:demo)

using OceanRobots
buoy=read(NOAAbuoy_monthly(),44013)
plot(buoy;option=:demo)

plot(b::SeaLevelAnomaly; dates=[], kwargs...)

using OceanRobots
sla=read(SeaLevelAnomaly(),:sla_podaac)
plot(sla)

plot(x::ArgoFloat; option=:standard, markersize=2,pol=Any[])

using OceanRobots, ArgoData, CairoMakie

argo=read(ArgoFloat),wmo=2900668)

f1=plot(argo,option=:samples)
f2=plot(argo,option=:TS)
f3=plot(argo,option=:standard)

plot(x::Gliders,ID)

using OceanRobots
lst=GDP.list_files()
drifter=read(SurfaceDrifter(),1,list_files=lst)
plot(drifter)

Base.read — Function

GliderFiles.read(x::Gliders, file::String)

Read a Spray Glider file.

read(x::ArgoFloat,wmo=2900668,files_list=GDAC.files_list())

using OceanRobots, ArgoData

wmo=2900668
lst=GDAC.files_list()
read(ArgoFloat(),wmo=2900668,files_list=lst)

read(x::SurfaceDrifter,ii::Int; list_files=[])

Open file list_files.filename[ii] from NOAA ftp server using NCDatasets.Dataset.

ftp://ftp.aoml.noaa.gov/pub/phod/lumpkin/hourly/v2.00/netcdf/ or the corresponding webpage

lst=GDP.list_files()
sd=read(SurfaceDrifter(),1,list_files=lst)

read(x::CloudDrift, file)

Read a GDP/CloudDrift file.

read(x::NOAAbuoy,args...)

Read a NOAA buoy file (past month).

read(x::NOAAbuoy_monthly,args...)

Read a NOAA buoy file (historical).

read(x::Gliders, file::String)

Read a Spray Glider file.

Internals

OceanRobots.OceanOPS.get_list — Function

get_list(nam=:Argo; status="OPERATIONAL")

Get list of platform IDs from OceanOPS API.

For more information see

https://www.ocean-ops.org/api/1/help/

https://www.ocean-ops.org/api/1/help/?param=platformstatus

list_Argo1=OceanOPS.get_list(:Argo,status="OPERATIONAL")
list_Argo2=OceanOPS.get_list(:Argo,status="CONFIRMED")
list_Argo3=OceanOPS.get_list(:Argo,status="REGISTERED")
list_Argo4=OceanOPS.get_list(:Argo,status="INACTIVE")

OceanRobots.OceanOPS.get_list_pos — Function

get_list_pos(nam=:Argo; status="OPERATIONAL")

Get list of platform positions from OceanOPS API.

For more information see

https://www.ocean-ops.org/api/1/help/

https://www.ocean-ops.org/api/1/help/?param=platformstatus

OceanRobots.OceanOPS.get_platform — Method

get_platform(i)

Get info on platform with id=i (e.g., float or drifter) from OceanOPS API.

For more information see https://www.ocean-ops.org/api/1/help/

list_Drifter=OceanOPS.get_list(:Drifter)
tmp=OceanOPS.get_platform(list_Drifter[1000])

OceanRobots.OceanOPS.get_url — Function

get_url(nam=:Argo; status="OPERATIONAL")

API/GET URL to OceanOPS API that will list platforms of chosen type.

Two URLs are reported; the second includes platform positions.

For more information see

https://www.ocean-ops.org/api/1/help/

https://www.ocean-ops.org/api/1/help/?param=platformstatus

https://www.ocean-ops.org/api/1/help/?param=platformtype

OceanRobots.OceanOPS.list_platform_types — Method

list_platform_types()

List platform types.

Base.read — Method

read(x::SurfaceDrifter,ii::Int; list_files=[])

Open file list_files.filename[ii] from NOAA ftp server using NCDatasets.Dataset.

ftp://ftp.aoml.noaa.gov/pub/phod/lumpkin/hourly/v2.00/netcdf/ or the corresponding webpage

lst=GDP.list_files()
sd=read(SurfaceDrifter(),1,list_files=lst)

OceanRobots.GDP.download — Function

download(list_files,ii=1)

Download one drifter file from NOAA ftp server.

ftp://ftp.aoml.noaa.gov/pub/phod/lumpkin/hourly/v2.00/netcdf/

list_files=GDP.list_files()
fil=GDP.download(list_files,1)

OceanRobots.GDP.list_files — Method

list_files()

Get list of drifter files from NOAA ftp server or the corresponding webpage.

ftp://ftp.aoml.noaa.gov/pub/phod/lumpkin/hourly/v2.00/netcdf/

https://www.aoml.noaa.gov/ftp/pub/phod/lumpkin/hourly/v2.00/netcdf/

Base.read — Method

read(x::CloudDrift, file)

Read a GDP/CloudDrift file.

OceanRobots.GDP_CloudDrift.add_ID! — Method

add_ID!(df,ds)

OceanRobots.GDP_CloudDrift.add_index! — Method

add_index!(df)

OceanRobots.GDP_CloudDrift.region_subset — Method

region_subset(df,lons,lats,dates)

Subset of df that's within specified date and position ranges.

OceanRobots.GDP_CloudDrift.to_DataFrame — Method

to_DataFrame(ds)

OceanRobots.GDP_CloudDrift.to_Grid — Method

to_Grid(gdf,grid)

OceanRobots.GDP_CloudDrift.trajectory_stats — Method

trajectory_stats(gdf)

Base.read — Method

read(x::NOAAbuoy,args...)

Read a NOAA buoy file (past month).

Base.read — Method

read(x::NOAAbuoy_monthly,args...)

Read a NOAA buoy file (historical).

OceanRobots.NOAA.download — Function

NOAA.download(stations::Union(Array,Int),path=tempdir())

Download files listed in stations from ndbc.noaa.gov to path.

OceanRobots.NOAA.download — Method

NOAA.download(MC::ModelConfig)

Download files listed in MC.inputs["stations"] from ndbc.noaa.gov to pathof(MC).

OceanRobots.NOAA.download_historical_txt — Method

NOAA.download_historical_txt(ID,years)

Download files from https://www.ndbc.noaa.gov to temporary folder for chosen float ID and years.

OceanRobots.NOAA.read_historical_monthly — Function

NOAA.read_historical_monthly(ID,years)

Read files from https://www.ndbc.noaa.gov to temporary folder for chosen float ID and year y.

OceanRobots.NOAA.read_historical_nc — Method

NOAA.read_historical_nc(ID,year)

Read files from https://www.ndbc.noaa.gov to temporary folder for chosen float ID and year y.

OceanRobots.NOAA.read_historical_txt — Method

NOAA.read_historical_txt(ID,y)

Read files from https://www.ndbc.noaa.gov to temporary folder for chosen float ID and year y.

OceanRobots.NOAA.read_station — Function

NOAA.read_station(station,path=tempdir())

Read station file from specified path, and add meta-data (units and descriptions).

Base.read — Method

read(x::Gliders, file::String)

Read a Spray Glider file.

OceanRobots.ArgoFiles.download — Method

ArgoFiles.download(files_list,wmo)

Download an Argo profiler file.

OceanRobots.ArgoFiles.readfile — Method

ArgoFiles.readfile(fil)

Read an Argo profiler file.

OceanRobots.ArgoFiles.scan_txt — Function

ArgoFiles.scan_txt(fil="ar_index_global_prof.txt"; do_write=false)

Scan the Argo file lists and return summary tables in DataFrame format. Write to csv file if istrue(do_write).

ArgoFiles.scan_txt("ar_index_global_prof.txt",do_write=true)
ArgoFiles.scan_txt("argo_synthetic-profile_index.txt",do_write=true)

Base.read — Method

GliderFiles.read(x::Gliders, file::String)

Read a Spray Glider file.

OceanRobots.OceanSites.index — Method

index()

Download, read and process the oceansites_index.txt file. Return a DataFrame.

oceansites_index=OceanSites.index()

OceanRobots.OceanSites.read_WHOTS — Function

read_WHOTS(fil)

Read an WHOTS file.

file="DATA_GRIDDED/WHOTS/OS_WHOTS_200408-201809_D_MLTS-1H.nc"
data,units=OceanSites.read_WHOTS(file)

OceanRobots.OceanSites.read_variables — Method

read_variables(file,args...)

Open file from opendap server.

file="DATA_GRIDDED/WHOTS/OS_WHOTS_200408-201809_D_MLTS-1H.nc"
OceanSites.read_variables(file,:lon,:lat,:time,:TEMP)

OceanRobots.podaac_sla.subset — Method

podaac_sla.subset()

For download directions, see this site

podaac_sla.subset()

OceanRobots.cmems_sla.subset — Method

cmems_sla.subset()

For download directions, see this site

For data documentation, see this page

cmems_sla.subset(username=username,password=password)